pyproject.toml
[project]
name = "chatproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chainlit>=2.3.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
